% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netReducer.R
\docType{package}
\name{netReducer-package}
\alias{netReducer}
\alias{netReducer-package}
\title{netReducer: Analysing the impact of sample size and species composition in network analysis}
\description{
The package takes your raw matrices, and repeatedly analyses a given subset of nodes, so as to show how
metrics vary depending on the raw data input
}
\details{
netReducer currently consists of only one basic command: \code{\link{netreducing}} for comparing how metrics vary
depending on how many nodes are used and their identity, and \code{\link{shrink}}, which is used by  \code{\link{netreducing}}.
No example data is yet included, see the readme at https://github.com/hemprichbennett/netReducer for examples.
}
\author{
\strong{Maintainer}: David Hemprich-Bennett \email{hemprich.bennett@gmail.com}

}
\keyword{internal}
