% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{calculate variance in network level metrics caused by sample size used}
\usage{
shrink(mat, n_cols, itval = NA, netname = NA, metric,
  metric_type = "network", network_level = NA, collapse = T)
}
\value{
Sends values to netreducing
}
\description{
Used by \code{\link{netreducing}}
}
\examples{
pending


}
