% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netreducing.R
\name{netreducing}
\alias{netreducing}
\title{calculate variance in network level metrics caused by sample size used}
\usage{
netreducing(input, input_type, n_iterations = 100, min_nodes, metric_chosen,
  type_chosen, level = NA)
}
\arguments{
\item{input}{Your input}
}
\value{
Produces a dataframe showing the changes in a given metric when varying the number of nodes in your matrix
}
\description{
Acts as a wrapper for the \code{\link{shrink}} function
}
\examples{
pending
}
